Fix int8 pixel overflow corrupting OpenCL 4PCF parity
Third instance of the int8 direction-pixel overflow (after the
non-periodic Fortran pass, fixed in v2.5.0, and the periodic pass, fixed
two commits ago): the OpenCL CSR packed csr_phi as int8 and the parity
kernel declared phi as signed char, so on the default 8x32 = 256-pixel
grid any index above 127 wrapped negative and indexed the host-built
chirality sign table out of bounds -- src_opencl/validate.sh reported
4PCFp relative errors of ~1e2 (reproduced on the pre-change baseline;
present since the v2.5.0 default-grid bump, which the 4x16 = 64-pixel
grid that fit int8 predates).

csr_phi is now int16, the kernel takes __global const short *phi, and
cl_env_module gains a cl_buf_in_i16 upload helper. The OpenACC CSR was
already int16; -exactparity never used pixels.

Verified on an Apple M1: src_opencl/validate.sh passes all five modes,
with 4PCFp at 1.4e-7 vs the double-precision CPU reference (identical
precision to the even channel, as expected -- the sign table keeps the
parity sign bit-exact and only the fp32 weight magnitude differs), both
plain and with GRAMSCI_CL_FORCE_TILED=1 exercising the watchdog path.

+++ src_opencl/csr_cl_module.F90
@@ -9,7 +9,9 @@
 !   csr_ptr(i) .. csr_ptr(i+1)-1   — range of edge indices for node i (1-based)
 !   csr_id(e)                       — neighbor node index for edge e
 !   csr_dist(e)                     — distance-bin index (int8)
-!   csr_phi(e)                      — direction-pixel index (int8, 4PCFp only)
+!   csr_phi(e)                      — direction-pixel index (int16, 4PCFp only:
+!                                     up to ntheta*nphi = 32767 pixels, 256 on
+!                                     the default 8x32 grid — too big for int8)
 !
 ! Each jagged row is freed as it is copied so peak host RAM stays near one
 ! graph copy.  csr_ptr / csr_total_edges are 64-bit (graphs may exceed 2^31
@@ -19,14 +21,14 @@
 !   call deallocate_csr()      at cleanup
 ! ---------------------------------------------------------------------------
 module csr_cl_module
-  use iso_fortran_env, only: int8, int32, int64
+  use iso_fortran_env, only: int8, int16, int32, int64
   use config_module
   implicit none
 
   integer(int64), allocatable :: csr_ptr(:)  ! shape (N+1), 1-based offsets
   integer(int32), allocatable :: csr_id(:)   ! shape (total_edges)
   integer(int8),  allocatable :: csr_dist(:) ! shape (total_edges)
-  integer(int8),  allocatable :: csr_phi(:)  ! shape (total_edges), 4PCFp only
+  integer(int16), allocatable :: csr_phi(:)  ! shape (total_edges), 4PCFp only
   integer(int64) :: csr_total_edges = 0
 
 contains
